Implement Drepper's improvements to scratch::mutex.
This new version of `mutex::lock()` is equivalent to version `mutex3`
as presented in Ulrich Drepper's "Futexes are Tricky", page 8. It's
a better implementation because, in the contended case, the old code
would do two CMPXCHG instructions per wait, and the new code does only
one XCHG instruction per wait.