Add an implementation of std::lock_guard.
Creating a member variable of reference type means that the copy and
move operations for this class will be implicitly deleted, so there's
no need to delete them explicitly (except for quality-of-error-messages).

The variable `adopt_lock` is marked `inline` in C++17, which is a good idea
in general, but here I'm leaving off the `inline` specifier so that this
code will still compile in C++14. (`inline` variables are new in C++17.)