Add an implementation of std::condition_variable.
This is based on code from a blog post by Rémi Denis-Courmont:
https://www.remlab.net/op/futex-misc.shtml

Notice that `scratch::condition_variable` and `scratch::mutex` both use
futexes, but `scratch::condition_variable` doesn't exploit this
coincidence in any way.

Notice also that if there are a lot of threads waiting on a condition
variable, calling `notify_all` on that condition variable is going to
cause all those threads to stampede from "waiting on `futex_wait(m_value)`"
to "waiting on `lk.lock()`". If all these locks `lk` are associated with
the same mutex, then they'll just queue up again over there. (This is
likely, in practice.)

But it's actually possible for all those locks to be associated with
*different* mutexes:

    condition_variable_any cv;
    mutex m1, m2, m3;
    std::thread t1([]{ unique_lock<mutex> lk(m1); cv.wait(lk1); });
    std::thread t2([]{ unique_lock<mutex> lk(m2); cv.wait(lk2); });
    std::thread t3([]{ unique_lock<mutex> lk(m3); cv.wait(lk3); });
    cv.notify_all();

So there's not very much that a Standard-conforming implementation
can do to make `notify_all` more efficient.