Add an implementation of `std::is_base_of<Base, Derived>`.
This code is based on the implementation from cppreference.com.

The essential idea here is that we have two overloads of
`is_base_of_test<B>`: one that's a good match for pointers to
derived classes of `B`, and one that's a less-good match for
pointers to any object type at all.  So the expression
`decltype(is_base_of_test<B>(declval<D*>()))` evaluates to
`true_type` if `B` is a public base class of `D`, and
to `false_type` if `B` and `D` are unrelated.

However, that doesn't cover all the possibilities! If `B` is
a *private* base class of `D`, then overload resolution selects
the first overload, but it turns out to be inaccessible, and
so SFINAE kicks in. It is in *this* situation --- a private
base class --- that we use the base template `is_base_of_impl`.