Add a Python script to check the #include structure.
By default, the script will spew a bunch of output, and raise a RuntimeError
with an appropriate message if it detects a missing file, a `std::` identifier
that's not on the small built-in whitelist, or a violation of include-what-you-use.

    ./*/generate-dependency-graph.py -I include --root include

With the `--dot` option, the script will output a GraphViz file suitable for
passing to the `dot` utility.

    ./*/generate-dependency-graph.py -I include --root include --dot | dot -Tpdf > graph.pdf