Add implementations of all the add_foo<T> and remove_foo<T> type traits.
Notice that the `add_foo<T>` traits generally follow the format "add a
foo qualifier to `T`, unless the resulting construct would be forbidden
by the language, in which case leave `T` alone."  We can express this
idea in a straightforward manner via SFINAE: we just get into a SFINAE
context and try matching `T` against `remove_foo_t<foo T>`.

In the case of `add_const`, `add_volatile`, and `add_cv`, we don't even
need to apply SFINAE, because the language permits adding those qualifiers
to all types, even when the result is just `T` again (for example,
`int& const` is the same type as `int&`; see also "is-function.md").