Add a non-standard "open addressed" hash-set.
`scratch::robin_hood_set<T>` is a completely original implementation
of the algorithms described in this blog post by Sebastian Sylvan:
https://www.sebastiansylvan.com/post/robin-hood-hashing-should-be-your-default-hash-table-implementation/

We use `optional<T>` where possible, in the belief that
`scratch::tombstone_traits` can make it as efficient as (or more efficient
than) the manual bit-twiddling of Sylvan's implementation. On the other
hand, we do actually still need to twiddle one bit; we copy Sylvan's
approach of treating a hash value of `0` as a special case.