Implement allocator<T>, allocator_traits, and pointer_traits.
We need `allocator<T>` before we can implement `vector`, and we
can't really use `allocator<T>` properly without `allocator_traits`.
`allocator_traits` factors out some common code into `pointer_traits`
(which is really just a helper for `allocator_traits`). So these
three pieces of the library pretty much come as a single unit.

Notice that `allocator_traits` provides static methods for constructing
and destroying objects of type T (and in this implementation but not in
the Standard, also for default-constructing them). This is provided by
the Standard solely as a means of implementing `scoped_allocator_adaptor`.
The instance method `scoped_allocator_adaptor::construct(p, args...)` is
implemented as a placement-new expression constructing `T(args..., *this)`.

Yes, this is a little bit convoluted; but it's actually kind of sane,
if your goal is to make something like `scoped_allocator_adaptor` possible.

Also notice that the `construct` and `destroy` (and `default_construct`)
methods are actually templated on `class U`, so that you can use an
allocator of any type to construct or destroy an object of any type.
This is mandated by the Standard, apparently to reduce the number of
distinct specializations `allocator<T>` required by a sufficiently clever
container type.

These methods are `constexpr` in my library but non-`constexpr` in the
actual Standard.