Implement std::move_iterator<It>.
This implementation differs from the standard's in that I didn't bother to
implement `make_move_iterator` because we're in C++17 now and have constructor
template type deduction.

The reason `pointer` is a synonym for `It`, instead of for
`iterator_traits<It>::pointer`, is because the former choice makes it easy
to express `operator->()` as simply returning `base()`. If we made the
latter choice, we'd have to make `operator->()` return one of the following
expressions:

- `base()`, which fails to compile if `base()` is not a pointer
- `base().operator->()`, which fails to compile if `base()` *is* a pointer
- `&*base()`, which fails to compile if `*base()` is an rvalue

So it's simplest to just return `base()` and let the caller sort out whether
it needs to call `operator->()` on that value recursively. (This is the same
choice the Standard made; I'm just explaining the rationale because it was
not at all obvious to me originally.)