Implement "tombstone_traits".
This is an idea that came out of Mark Zeren's talk about strings
at C++Now. The basic idea is for otherwise opaque classes to expose
a sort of side-channel by which their otherwise unused bits can
be exploited for size optimizations.

After this patch, `sizeof(optional<bool>) == 1`; in fact,
`sizeof(optional<optional<optional<... T>>>) <= sizeof(optional<T>)`
as long as there are fewer than 254 nested "optional"s there,
because even if we can't fold the `m_has_value` bool into the
padding bits of `T`, we can certainly fold the next 254 bools
into the unused bits of that first `m_has_value`!

This trick may or may not exploit undefined behavior. I think it
does not, although it might certainly be tempting for programmers
to accidentally use undefined behavior in their implementations
of `index` and `set_spare_representation`.

This trick does certainly break encapsulation, in that
`tombstone_traits<T>` needs access to the layout of class `T`
so that it knows which bits are safe to trample on.

This trick also breaks the constexpr-ness of `optional`'s
default constructor, in that it now unavoidably involves a
`reinterpret_cast`. I don't know how to get around that.