Initial implementation of `std::variant`.
Notice that this version of `std::visit` can handle only a single variant;
it's not variadic (yet). Also notice that it can handle the valueless_by_exception
case without an extra branch; the throwing case is just rolled into the array
of possible actions with the rest of the cases.

...However, by doing this, we lose the ability to take visitors that return
non-void types! This seems to be a problem.