Emulate Clang bug 33425.
If we're trying to cast subobject A to type B, where A is a subobject
of some subobject of type B, `dynamic_cast<B>(a)` should give us
that B subobject... unless the inheritance path from B to A is
non-public.

However, even if the inheritance path is non-public, if there *also*
exists a public path from B down to the root, and a public path from
the root back up to A, then casting from A to B should work.

libc++abi's `__dynamic_cast` gets this wrong in some cases.

https://bugs.llvm.org/show_bug.cgi?id=33425