Emulate Clang bug 33439.
If we're trying to cast subobject A to type B, where A and B are
unrelated types, `dynamic_cast<B>(a)` should cast down to the root
(the most derived object, that is) and then back up to B, if possible.
If B is an ambiguous base type of the root, then this won't be
possible and `dynamic_cast<B>(a)` should fail.

However, libc++abi's `__dynamic_cast` gets this wrong whenever
type B has direct bases of its own (that is, whenever B is not
a leaf). In that case, casting to ambiguous base B will actually
return the address of the first B subobject in inheritance-graph
order.

Notice that inheritance-graph order is not the same as layout order.

https://bugs.llvm.org/show_bug.cgi?id=33439