Implement `std::void_t`, and the non-standard `std::false_v`.
This "metafunction" utility can be used similarly to `decltype((void)(...))`
in expression SFINAE. To see an example of `void_t`'s use, look at "iterator-traits.h".

The utility `false_v<T>` can be used similarly to `sizeof(T)==0` in
static-asserts, when we want to `static_assert(false)` any time a particular
specialization is instantiated, but we don't want the compiler to eagerly
halt compilation just because the static-assert condition is a non-dependent
expression. To see an example of `false_v`'s use, look at "pointer-traits.h".