Add a "counted copying iterator".
Most STL containers have a method `assign(size_t, const T&)` and another
method `assign(It, It)` that works with input iterators. By creating an
iterator that yields exactly `n` copies of a `const T&`, we can make it
really easy to implement the former in terms of the latter.