Make `dynamicast` also work for the 64-bit MSVC ABI, and compile on MSVC.
MSVC doesn't support C++14 variable templates or C++03 expression SFINAE,
so we have to fall back on overload resolution in order to get things to
come out correctly. (In Microsoft's case, "SFINAE failure is not an error",
so if you have a chain of SFINAE functions like my `dynamicast-cxx14.h`,
MSVC will just silently choose the last one in the chain, regardless of
the SFINAE conditions on the earlier, better-matching ones.)

This finds some `dynamic_cast` bugs in the MSVC runtime, of course.