Add an implementation of `observer_ptr<T>`, the simplest smart pointer.
I'm going to try basing all of the library's smart pointers on a class
I'm calling `smart_ptr_base<T>`, whose purpose is to centralize the
metaprogramming for `ptr<T>` versus `ptr<T[]>` versus `ptr<void>`.
It remains to be seen if this will actually be useful for `unique_ptr`
and `shared_ptr`.

Notice that I did not implement the comparison operators < <= > >=
for `observer_ptr`; allowing them to be directly comparable strikes
me as a bad idea, and besides which, N4282 requires the ordering to
be a total ordering a la `std::less<void*>`, and I haven't implemented
`std::less` yet.