Add an implementation of the non-standard `guarded<T>`.
This class template comes from [libGuarded](https://github.com/copperspice/libguarded).
I've spiced it up a bit with a method to extract the underlying
`unique_lock<M>` (since I see no other way to use it with a
condition variable), and to check that the lock exists before
dereferencing in `operator->` and `operator*`.

I've attempted to use `shared_ptr<guarded<shared_state>>` to implement
`promise` and `future`, but even with the addition of the lock-extraction
function, it still seems hard to `wait()` on the internals, and/or to access
them locklessly once the waiting is over.