Add a non-standard example of a (standard) "fancy pointer".
The one possible problem with this fancy pointer is that it does not
participate in `iterator_traits`. libstdc++'s `std::vector::resize`
requires that its fancy pointers also be iterators; but I think that
that's a bug in libstdc++.

It's unclear to me whether `static_cast<T*>(fancy_ptr)` needs to work,
or whether allocator-aware containers should always be using
`std::addressof(*fancy_ptr)` to convert fancy pointers into raw pointers.
The latter invokes undefined behavior if `fancy_ptr` is null, so I'm
sticking with the `static_cast` as far as this library is concerned.