Add a non-standard example of a (standard-ish) "fancy pointer".
The one possible problem with this fancy pointer is that it does not
participate in `iterator_traits`. The standard requires that fancy
pointer types also by iterators; this is used in the implementation
of libstdc++'s `std::vector::resize`, although it needn't be.

It's unclear to me whether `static_cast<T*>(fancy_ptr)` needs to work,
or whether allocator-aware containers should always be using
`std::addressof(*fancy_ptr)` to convert fancy pointers into raw pointers.
The latter invokes undefined behavior if `fancy_ptr` is null, so I'm
sticking with the `static_cast` as far as this library is concerned.