Add a "shared_polymorphic_allocator" that decides for itself whether to propagate.
This doesn't actually work, because the container's own code
(e.g. `std::vector::operator=`) assumes that allocator assignment never fails.
By the time we reach the move-assignment of allocator to allocator, we must
know for certain whether the assignment is going to succeed or fail (that is,
successfully transfer the rhs's old value to the lhs, or be a no-op).
For example:

    vector& operator=(vector&& rhs) noexcept {
        if constexpr (Alloc_traits::propagate_on_container_move_assignment_v) {  /* !! */
            // we can adopt the new allocator and its memory wholesale
            this->clear();
            this->m_allocator() = std::move(rhs.m_allocator());
            this->m_data() = std::exchange(rhs.m_data(), pointer{});
        } else if (m_allocator() == rhs.m_allocator()) {
            // since the allocator is unchanged, we can adopt the memory
            this->clear();
            this->m_data() = std::exchange(rhs.m_data(), pointer{});
        } else {
            // we must not propagate this new allocator and thus cannot adopt its memory
            this->assign(rhs.begin(), rhs.end());
        }
    }

For this "optionally reseatable" allocator type to work out, we'd need
the line marked `/* !! */` to look more like this:

        if (this->m_allocator().is_reseatable()) {  /* !! */