Implement `list<T>` as an allocator-aware container; and add `pointer_traits::pointer_to`.
Notice the extreme similarity between `list<T>` and `forward_list<T>`.

The major difference between `list` and `forward_list` is that whereas
the latter's "end" iterator is represented by a null pointer, the
former's needs a unique representation because it needs to be decrementable.

In `forward_list`, the representation of iterator `x.before_begin()`
includes a raw pointer pointing at `&x.m_before_head()`.

In `list`, the representation of iterator `x.end()`
includes a raw pointer pointing at `&x.m_after_tail()`. However ---
and this is the key point --- the *list node* representation of
`prev(x.end())->next` actually stores a *fancy pointer* pointing
at `&x.m_after_tail()`!  This implies that we need a way to convert
the raw, non-allocated pointer `&x.m_after_tail()` into a fancy pointer
suitable for comparison and dereferencing (but not deallocation).

Converting a raw pointer into a (non-deallocatable) fancy pointer
is a new primitive operation that needs to be delegated to the
`pointer_traits` traits class. The standard name for this function
is `pointer_traits<P>::pointer_to(T& r)`. Notice that this function
doesn't work if what we need is a null pointer, or a pointer to
one-past-the-end-of-an-array; this is probably short-sighted of the
standard version, but since this trait is used only by `std::list`
and a few other node-based containers, the omission is not problematic
at the moment.