Implement `scoped_allocator_adaptor`, more or less.
This implementation doesn't include the five special overloads for `std::pair`,
but otherwise I think it's correct. (The special overloads for `std::pair` are
unnecessary in C++17 with the introduction of `allocator_arg`, although the
standard library has to preserve them for backwards compatibility.)

We omit the inheritance (IS-A) relationship between `scoped_allocator_adaptor<A, ...>`
and `A`, since in general a `scoped_allocator_adaptor` is *not* Liskov-substitutable
for an instance of its `outer_allocator` type.