Implement `std::string` as an allocator-aware type.
This implementation is not yet parameterized on `char_traits<T>`.

Notice that even though we are an allocator-aware container (well,
not technically a container, but basically), we do not use
`allocator_traits<A>::construct` to construct our array of `T`.
This is permissible because `T` is required to be a POD type, and
thus it must have a trivial constructor. It's still not clear to me
how the Standard wording is meant to imply the lack of `construct`
and `destroy` calls; but we definitely can't call them, because that
would result in calling user-defined code (which might throw) in
lots of places that cannot reasonably be allowed to throw if we
want to preserve the strong exception guarantee. And `string`
--- unlike `vector` --- is explicitly required to preserve the
strong exception guarantee.