Implement a version of `boost::interprocess::offset_ptr<T>`.
This type is just like `T*`, except that its bit-level representation is
"position-independent": the pointer is stored as the integer difference
between the referred-to object and `this`.

The non-standard `offset_allocator_adaptor<A>` is an allocator type
that behaves just like `A`, except that its member typedef `pointer`
is an alias for `offset_ptr<A::value_type>` instead of for `A::pointer`.
This has the effect of erasing any "fancyness" about `A::pointer`,
so we assert that `A::pointer` and `A::value_type*` are interconvertible
(and thus that any "fancyness" about `A::pointer` is irrelevant to its
behavior, e.g. to its deallocatability).

Notice that `offset_allocator_adaptor<A>` contains a member of type `A`.
Any raw pointers in the object representation of `A` itself will not
magically replace themselves with `offset_ptr`.