Add atomic_shared_ptr<T> and atomic_unique_ptr<T>.
`atomic_unique_ptr<T>` was proposed in N4162, but in a silly way that
wasn't actually implementable (e.g. it had a const `load()` method
that returned a *copy* of the `unique_ptr` value stored inside it).
This version is more sane. I don't think there's much of a use-case
for `atomic_unique_ptr<T>`, though.

`atomic_shared_ptr<T>` was proposed in N4162, and has continued to
be active; the most recent proposal for it is Alisdair Meredith's
P0718 "Revising atomic_shared_ptr for C++20".

There is a "lock-free" (read: perhaps buggy) implementation here:
https://bitbucket.org/anthonyw/atomic_shared_ptr/src/

The implementation in this repo is *not* lock-free, but holds a lock
only for as long as it takes to swap some pointers; we never hold the
lock during:
- user-defined destructor code (e.g. `~T()`)
- user-defined constructor code (e.g. `Deleter(Deleter&&)`)
- allocation (such as the creation of a new control block)
- deallocation (such as the destruction of a control block)
The operations we allow under the lock are:
- shared_ptr swap
- shared_ptr copy (increment refcount, but never decrement it)