Add some pieces of the <random> header.
`linear_congruential_engine` is broken for large `m != 0`; it doesn't
do the modulus correctly when the intermediate result overflows.

`random_device` is broken in that a short read (due to `EINTR`) can
cause it to throw an exception instead of completing successfully.

`random_device` implements the `rd.generate(from, to)` member function
required by P0205's `SeedGenerator` concept:
http://www.open-std.org/jtc1/sc22/wg21/docs/papers/2016/p0205r0.html

This means that you can seed the ISAAC generator, for example, with

    random_device rd;
    isaac g(rd);

The type trait `scratch::is_seed_generator_v<G>` yields `true` whenever
`declval<G>().generate(p, q)` is a valid expression. Notice that the
value-category of `G` is preserved; so for example it is possible that
`is_seed_generator_v<Foo&> && !is_seed_generator_v<const Foo&>`.
Because UniformRandomBitsEngine's constructor and `seed()` function both
perfectly forward the seed-generator argument, it is even possible to
write this:

    isaac g(random_device{});
    g.seed(random_device{});

This is currently impossible in C++17 because the overload of `seed()`
that takes a seed-sequence `Sseq&` takes the sequence by lvalue reference.

`isaac_engine` and `isaac_fast_engine` (a.k.a. `isaac`) are equivalent
implementations of the ISAAC cryptographically secure PRNG;
see http://burtleburtle.net/bob/rand/isaac.html

`seed_seq` directly implements the algorithm described in the Standard.
I have confirmed that it produces the same output as libc++ and libstdc++.