Implement `call_once` and `once_flag`.
Notice that `call_once` is essentially just a member function of
`once_flag`, so that's how I've implemented it (and a free function
that just calls the member-function version).

This implementation is based on libc++. Notice that it is *not* the implementation
you might think of at first, using an atomic flag and compare-exchange; that's
because `once_flag` does not count simply "has this line been hit" but actually
"has this line been hit and not thrown an exception"! Anyone entering the `call_once`
while the callback is executing must *stop and wait* until the callback has completed,
and, if the callback threw, the waiter might have to step up and finish the job
itself.

Because some threads must *stop and wait*, we deduce that we must use a condition variable
and a mutex. libc++ uses a global (static) mutex and condition variable; libpthread
uses a non-static member mutex. I take the latter approach because it's simpler.