Implement std::system_error.
Notice that std::system_error depends on std::string, which means that
it cannot be nothrow-copyable. If it didn't have that pointless dependency,
i.e. if it really just contained a std::error_code and nothing else, then
it could have been trivially copyable!

The rationale for making it so heavyweight is presumably that you're
going to be throwing this object, i.e. not copying it (since a throw-expression
will usually result in a move, not a copy, of the thrown object), and
the cost of the memory allocation for std::string (if it is not SSO'ed)
will be negligible compared to the cost of the throw itself.