Implement `reftuple` as a "tuple-of-references" type.
In this library, `scratch::tuple<Ts..>` works only with object types,
and `scratch::reftuple<Ts...>` works only with reference types, and
never the twain shall meet.

The `reftuple` type is intended specifically to make `forward_as_tuple`
work for metaprogramming, and to satisfy the "least privilege principle"
otherwise. Therefore it is not default-constructible, not assignable,
and not swappable. The non-assignability means that if I were to implement
a `tie` returning `reftuple`, it wouldn't emulate multi-assignment the
way `std::tie` does in standard C++. I'm not sure yet if this is a
feature or a bug.

This `reftuple` business is completely non-conforming with the C++ standard,
of course. The most obvious breakage is if you try to do this:

    template<class... Refs>
    void foo(scratch::tuple<Refs...> t);

    ... foo(scratch::forward_as_tuple(x,y,z); ...

An appropriate workaround would be to template on the "tuple-like"
class template itself:

    template<template<class...> class Tuple, class... Refs>
    void foo(Tuple<Refs...> t);

    ... foo(scratch::forward_as_tuple(x,y,z); ...