Implement `constexpr_allocator` and `inplace_allocator`.
This implementation of `constexpr_allocator` is taken largely from
https://github.com/ZaMaZaN4iK/constexpr_allocator/blob/master/test.cpp

The type-trait `scratch::is_constexpr_allocator<A>` is provided as
a customization point for user-defined class types that want to
act like `constexpr_allocator` (that is, to signal to containers
that `destroy` and `deallocate` are not provided and must not be
called).

The implementation of `inplace_allocator` is my own. This represents
an "allocator" with no `allocate` and `deallocate` methods; it is
intended to be passed to `basic_any` to indicate that we want SFINAE
instead of heap-allocation in the "object too big for buffer" case.

The type-trait `scratch::is_inplace_allocator<A>` is provided as
a customization point for user-defined class types that want to
act like `inplace_allocator` (that is, to signal to containers that
`allocate` and `deallocate` are not provided and must not be called).