Change the condition for "simply relocating" to match reality instead of Standardese.
In reality, the move-constructibility of T is not a sufficient condition
to "simply relocate"; we must first be 100% sure that the allocator isn't
going to throw any exceptions on its own.

Notice that we could also safely relocate (instead of copy + destroy) in
one more special case:

    Alloc_traits::template has_trivial_construct_and_destroy_v<T> && (
        is_trivially_destructible_v<T> && !is_copy_constructible_v<T>
    )

Normally relocation is unsafe because we don't want the caller to be able
to observe moved-from objects after a throw (so we must copy+destroy instead).
But if "copy" is impossible, then we will move+destroy and be okay with
the caller observing moved-from objects. We are *still* not okay, in that case,
with the caller observing *destroyed* objects! But if the destruction is
trivial, then a destroyed object looks just the same as a non-destroyed
object, and so relocation is okay again.

This special case requires a `T` with a trivial destructor, a deleted
copy-constructor, and a throwing move-constructor; therefore it is not
plausible enough to bother encoding in the library.